Binomial series fix

+++ IrrationalNumbers/IrrationalNumbers.Logic/ParameterNormalizationResult.cs
@@ -2,7 +2,7 @@
 {
     public class ParameterNormalizationResult
     {
-        public double NormalizedParameter { get; set; }
+        public BigDecimal NormalizedParameter { get; set; }
 
         public double FinalMultiplier { get; set; } = 1;
     }
